Fixed chained tweens.
When two tweens were chained to form an endless loop, they would execute properly only once. On second
time around they would only skip between the end values of each other. This was because the
```_object``` was overwritten in the ```update()``` and then used as the initialization value
in the next ```start()```.

Fixed by saving the initial values to a separate object and copying the initial values to ```_object```
in ```start()```.

As a bonus refactoring the ```_valuesEnd``` initialization to the ```to()``` avoids re-concatting the
```valuesEnd[property]``` array to itself with a new copy of the initial property in beginning on
every ```start()```.