Fix Tween synchronization when using repeat.
The current method of calculating repeat _startTime can cause Tweens to
fall out of sync. The current calculation is dependent on the update
time that triggers the repeat, not the original _startTime and _duration
of the Tween.  This issue can be readily observed by 1) creating two
Tweens with different durations, 2) setting them to repeat, and 3)
starting them at the same time. They slowing fall out of sync.

Fix: _originalStartTime, number of _repeats and _duration are used to
calculate a consistent repeat _startTime. _delayTime is also taken into
account.