`onStartEvery`, as `onStart` but called on every repeat
This is similar but not equivalent to `onRepeat`. The behaviour is as `onStart`, but it will trigger on every repeat. Unlike `onRepeat`, this means it will trigger at the start of the animation cycle, not the end of the previous. This is a critical difference if delays are in use. It also may give more expected results depending where in the draw cycle the TWEEN.update is called.